+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ReleaseDate
+
+### Added
+
+- Add ``repeat_interleave`` function
+
+### Fixed
+
+- Fix ``undo_shape_blocksparse`` function checking the incorrect dimensions
+
 ## [1.2.1] - 2024-09-25
 
 ### Changed